+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added trigger ```scrolling.uk.document``` after update grid layout
   - Added ARIA support to JavaScript core components
   - Fixed sticky wrapper resize on window resize/orientation change
+  - Fixed nestable+sortable: serialize only includes data attributes from HTML
 
 
 ### 2.17.0 (February 20, 2015)